fix: split cli_options.py/mcp_server.py under the 2000-line CI hard cap
Both files had drifted past the AI-readiness file-size hard limit,
which was failing the ai-readiness check and, transitively, every
unit-tests lane (test_ai_readiness.py asserts a clean tree) on main.

- Split the ADR-037 D10 CLI-contract metadata (FAMILY_FLAGS,
  COMPARE_FLAG_BUDGET*, count_visible_options, ...) out of
  cli_options.py into a new leaf module, cli_options_contract.py,
  re-exported for existing callers -- mirrors the existing
  cli_profiles.py split.
- Split the abi_estimate/abi_scan MCP tools out of mcp_server.py into
  a new sibling module, mcp_server_scan.py, imported for side-effect
  and re-exported -- mirrors the existing mcp_server_project.py split.
- Regenerated docs/reference/mcp-tools-reference.md and updated
  gen_mcp_reference.py's own text to name the new sibling module.
- Added abicheck.mcp_server.mcp_server_scan to the taint-purge lists
  in tests/test_mcp_reference.py and tests/test_cov95_misc.py, which
  temporarily mock the `mcp` package and restore a fixed list of
  abicheck.mcp_server sibling modules afterward -- the list predated
  mcp_server_scan and, left unpatched, let a later test observe a
  stale, mock-bound mcp_shared instance and silently skip a file-size
  check (only surfaced when running the full suite, not any one file).
- Added a mypy override entry for mcp_server_scan.py's @mcp.tool()
  decorators, matching mcp_server.py/mcp_server_project.py.

No behavior change. Full fast suite (23406 tests) now passes cleanly;
it previously had 2 failures on this branch's baseline (the file-size
violations) plus the 2 latent test-isolation failures this surfaced.
