+++ scripts/CLAUDE.md
@@ -31,7 +31,7 @@ Each must run with Python 3.10+ and the package installed in dev mode
 | `gen_detector_spec.py` | Regenerates the formal detector specification matrix (`docs/reference/detector-spec.{md,json}`) by fusing per-`ChangeKind` category (`checker_policy`), default verdict/severity/doc-slug (`policy_for`), and min evidence tier (`evidence_tiers`). `--check` gates sync; mirrored in `tests/test_detector_spec.py`. Run after adding a `ChangeKind`. | manual |
 | `gen_action_reference.py` | Regenerates the exhaustive GitHub Action inputs/outputs reference (`docs/reference/github-action-inputs.md`) from `action.yml`'s `description`/`required`/`default` fields — the single source of truth `docs/use/github-action.md`'s narrative previously restated by hand. `--check` gates sync; mirrored in `tests/test_action_reference.py`. Run after adding, renaming, or changing the default of an Action input/output. | manual |
 | `gen_cli_reference.py` | Regenerates the exhaustive CLI command/option reference (`docs/reference/cli-reference.md`) from the live Click command tree (`abicheck/cli.py` + registered sibling modules) — every command/subcommand/argument/option with its exact `help=` text, default, and choices. `--check` gates sync; mirrored in `tests/test_cli_reference.py`. Run after adding/renaming a command or option, or changing its `help=`/default. | manual |
-| `gen_mcp_reference.py` | Regenerates the exhaustive MCP tool parameter reference (`docs/reference/mcp-tools-reference.md`) from every `@mcp.tool()` function registered on the shared `mcp` instance — `abicheck/mcp_server.py`'s own tools plus `abicheck/mcp_server_project.py`'s sibling-module tools — via each function's signature and docstring `Args:` section. Requires the optional `mcp` extra to import. `--check` gates sync; mirrored in `tests/test_mcp_reference.py` (skipped without `mcp` installed). Run after adding/renaming an MCP tool or its parameters. | manual |
+| `gen_mcp_reference.py` | Regenerates the exhaustive MCP tool parameter reference (`docs/reference/mcp-tools-reference.md`) from every `@mcp.tool()` function registered on the shared `mcp` instance — `abicheck/mcp_server.py`'s own tools plus its sibling modules' (`abicheck/mcp_server_project.py`, `abicheck/mcp_server_scan.py`) tools — via each function's signature and docstring `Args:` section. Requires the optional `mcp` extra to import. `--check` gates sync; mirrored in `tests/test_mcp_reference.py` (skipped without `mcp` installed). Run after adding/renaming an MCP tool or its parameters. | manual |
 | `gen_python_api_reference.py` | Regenerates the exhaustive Python API reference (`docs/reference/python-api-reference.md`) from every name in `abicheck.service.__all__` — full function signatures and dataclass field lists. `--check` gates sync; mirrored in `tests/test_python_api_reference.py`. Run after changing `service.__all__` or a public function/dataclass's signature. | manual |
 | `gen_config_reference.py` | Regenerates the exhaustive `.abicheck.yml` key/type reference (`docs/reference/config-keys-reference.md`) from `abicheck.buildsource.inline.BuildConfig`'s strict-schema registries (`_KNOWN_TOP_KEYS`/`_KNOWN_BLOCK_KEYS`/`_BOOL_SUBKEYS`/`_STR_SUBKEYS`/`_LIST_SUBKEYS`). `--check` gates sync; mirrored in `tests/test_config_reference.py`. Run after adding/renaming a config key or changing its required type. | manual |
 | `gen_platform_matrix.py` | Regenerates the host-OS × binary-format capability matrix — `docs/reference/platforms.md`'s "Quick Reference: What Works Where" section, spliced between `<!-- BEGIN/END GENERATED: platform-matrix -->` sentinels — from `platform_capabilities.py`. `--check` gates sync; mirrored in `tests/test_platform_matrix.py`. Run after a symbol/type-diff capability or tooling requirement changes for a host/format pair (ADR-051 Stage 2). | manual |
